added order changed models for order

+++ Electro-goods-API/Models/Entities/OrderItem.cs
@@ -7,7 +7,7 @@
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public string? ProductNameUK { get; set; }
-        public int ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int Quantity { get; set; }
         public Order? Order { get; set; }
         public Product? Product { get; set; }
